requirements.txt
aiohttp==3.5.4
astroid==2.2.5
async-timeout==3.0.1
attrs==19.1.0
cachetools==3.1.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
discord.py==1.1.1
google-api-python-client==1.7.8
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.3.0
httplib2==0.12.1
idna==2.8
isort==4.3.17
lazy-object-proxy==1.3.1
mccabe==0.6.1
multidict==4.5.2
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pygsheets==2.0.1
pylint==2.3.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
typed-ast==1.3.1
uritemplate==3.0.0
urllib3==1.24.2
websockets==6.0
wrapt==1.11.1
yarl==1.3.0